solution

+++ 1373A.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main(){
 	int tc;cin>>tc;
 	while(tc--){
-		int a,b,c;
+		long long a,b,c;
 		cin>>a>>b>>c;
 		if (a>=c)
 		{
